requirements.txt
Django==5.2.4
djangorestframework==3.14.0
drf-yasg==1.21.7
requests==2.31.0
# mysqlclient==2.2.0  # MySQL 사용 시 주석 해제 